+++ Nos/Models/Event+CoreDataClass.swift
@@ -704,7 +704,20 @@ public class Event: NosManagedObject {
         })
     }
     
-    /// Returns the root event that this note is replying to, or nil if there isn't one.
+    var isReply: Bool {
+        rootNote() != nil
+    }
+    
+    /// Returns the event this note is directly replying to, or nil if there isn't one.
+    func referencedNote() -> Event? {
+        if let lastReference = eventReferences?.lastObject as? EventReference,
+            let referencedNote = lastReference.referencedEvent {
+            return referencedNote
+        }
+        return nil
+    }
+    
+    /// Returns the root event of the thread that this note is replying to, or nil if there isn't one.
     func rootNote() -> Event? {
         let rootReference = eventReferences?.first(where: {
             ($0 as? EventReference)?.marker ?? "" == "root"