+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Support likes
+- Show "posted" and "replied" headers on NoteCards
+- Navigate to replied to note when tapping on reply from outside thread view
+- Search by name on Discover view
 
 ## [0.1 (14)] - 2023-03-16Z
 - Nos now reads and writes your mutes the mute list shared with other Nostr apps.